requirements.txt
# Web Framework
Flask==3.0.2

# API and HTTP
requests==2.31.0

# Environment Variables
python-dotenv==1.0.1

# Machine Learning and Data Processing
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.5.1

# Other utilities
pickle-mixin==1.0.2
gunicorn
